requirements.txt

imageio==2.21.1
matplotlib==3.5.2
pandas==1.4.3                  
pretrainedmodels==0.7.4                  
streamlit==1.11.1
torch==1.12.1                  
torchvision==0.13.1                   
pillow==9.2.0
tqdm==4.64.0      
notebook==6.4.12      